Move charge and hvac commands to sub module
