+++ src/renault_api/cli/charge/__init__.py
@@ -1,0 +1,1 @@
+"""Renault CLI."""
