+++ src/renault_api/cli/charge/commands.py
@@ -1,0 +1,19 @@
+"""Command-line interface."""
+import click
+
+from . import control
+from . import history
+from . import schedule
+
+
+@click.group()
+def charge():
+    """Charge functionnality."""
+    pass
+
+
+charge.add_command(control.start)
+charge.add_command(control.mode)
+charge.add_command(history.history)
+charge.add_command(history.sessions)
+charge.add_command(schedule.schedule)
