+++ src/renault_api/cli/hvac/__init__.py
@@ -1,0 +1,1 @@
+"""Renault CLI."""
