+++ src/renault_api/cli/hvac/commands.py
@@ -1,0 +1,17 @@
+"""Command-line interface."""
+import click
+
+from . import control
+from . import history
+
+
+@click.group()
+def hvac():
+    """HVAC functionnality."""
+    pass
+
+
+hvac.add_command(control.start)
+hvac.add_command(control.cancel)
+hvac.add_command(history.history)
+hvac.add_command(history.sessions)
