+++ tests/cli/test_vehicle_charge.py
@@ -105,18 +105,19 @@ def test_sessions(mocked_responses: aioresponses, cli_runner: CliRunner) -> None
     assert expected_output == result.output
 
 
-def test_charging_settings_get(
+def test_charge_schedule_show_all(
     mocked_responses: aioresponses, cli_runner: CliRunner
 ) -> None:
     """It exits with a status code of zero."""
     initialise_credential_store(include_account_id=True, include_vin=True)
     fixtures.inject_get_charging_settings(mocked_responses)
 
-    result = cli_runner.invoke(__main__.main, "charge settings")
+    result = cli_runner.invoke(__main__.main, "charge schedule show")
     assert result.exit_code == 0, result.exception
 
     expected_output = (
         "Mode: scheduled\n"
+        "\n"
         "Schedule ID: 1 [Active]\n"
         "Day        Start time    End time    Duration\n"
         "---------  ------------  ----------  ----------\n"
@@ -127,7 +128,34 @@ def test_charging_settings_get(
         "Friday     13:15         13:30       0:15:00\n"
         "Saturday   13:30         14:00       0:30:00\n"
         "Sunday     13:45         14:30       0:45:00\n"
-        "\n\n"
+        "\n"
+        "Schedule ID: 2\n"
+        "Day        Start time    End time    Duration\n"
+        "---------  ------------  ----------  ----------\n"
+        "Monday     02:00         02:15       0:15:00\n"
+        "Tuesday    03:00         03:30       0:30:00\n"
+        "Wednesday  04:00         04:45       0:45:00\n"
+        "Thursday   05:00         06:00       1:00:00\n"
+        "Friday     06:00         07:15       1:15:00\n"
+        "Saturday   07:00         08:30       1:30:00\n"
+        "Sunday     08:00         09:45       1:45:00\n"
+    )
+    assert expected_output == result.output
+
+
+def test_charge_schedule_show_single(
+    mocked_responses: aioresponses, cli_runner: CliRunner
+) -> None:
+    """It exits with a status code of zero."""
+    initialise_credential_store(include_account_id=True, include_vin=True)
+    fixtures.inject_get_charging_settings(mocked_responses)
+
+    result = cli_runner.invoke(__main__.main, "charge schedule show --id 2")
+    assert result.exit_code == 0, result.exception
+
+    expected_output = (
+        "Mode: scheduled\n"
+        "\n"
         "Schedule ID: 2\n"
         "Day        Start time    End time    Duration\n"
         "---------  ------------  ----------  ----------\n"
@@ -154,7 +182,7 @@ def test_charging_settings_set(
     friday = "--friday T23:30Z,480"
     saturday = "--saturday 19:30,120"
     result = cli_runner.invoke(
-        __main__.main, f"charge settings --id 1 --set {monday} {friday} {saturday}"
+        __main__.main, f"charge schedule set --id 1 {monday} {friday} {saturday}"
     )
     assert result.exit_code == 0, result.exception
